pyproject.toml
[project]
name = "plugin-crosswalk"
version = "0.1.0"
description = "Cross-provider converter for agent skills and plugin packages"
readme = "README.md"
authors = [
    { name = "Max Ghenis", email = "mghenis@gmail.com" }
]
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.3",
]

[project.urls]
Homepage = "https://github.com/MaxGhenis/plugin-crosswalk"
Repository = "https://github.com/MaxGhenis/plugin-crosswalk"
Issues = "https://github.com/MaxGhenis/plugin-crosswalk/issues"

[project.scripts]
plugin-crosswalk = "plugin_crosswalk.cli:main"

[build-system]
requires = ["uv_build>=0.11.2,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
